# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_uthmanduro_BracketForge_internal_model.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  github_com_uthmanduro_BracketForge_internal_model.SubmitResultRequest:
    properties:
      sets:
        items:
          $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.SetScoreInput'
        minItems: 1
        type: array
    required:
    - sets
    type: object
  github_com_uthmanduro_BracketForge_internal_model.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  github_com_uthmanduro_BracketForge_internal_model.Tournament:
    properties:
      best_of:
        type: integer
      created_at:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      end_date:
        type: string
      format:
        $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.TournamentFormat'
      id:
        type: string
      name:
        type: string
      organization_id:
        type: string
      scoring_rules:
        items:
          type: integer
        type: array
      start_date:
        type: string
      status:
        $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.TournamentStatus'
      third_place_match:
        type: boolean
      updatedAt:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@bracketforge.io
    name: BracketForge Support
  description: Racket sports tournament management engine. Supports single elimination,
    round robin, and group + knockout formats with set-based scoring, three-way tiebreakers,
    BYE handling, and result correction.
  termsOfService: http://swagger.io/terms/
  title: BracketForge API
  version: 1.0.0
paths:
...
  /tournaments:
    get:
      description: Returns all tournaments belonging to the authenticated organisation
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.SuccessResponse'
            - properties:
                ' Message':
                  type: string
                Data:
                  properties:
                    tournaments:
                      items:
                        $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.Tournament'
                      type: array
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List tournaments
      tags:
      - tournaments
    post:
      consumes:
      - application/json
      description: 'Creates a new tournament. Format must be one of: single_elimination,
        round_robin, group_knockout. BestOf must be 3 or 5.'
      parameters:
      - description: Tournament configuration
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.CreateTournamentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.SuccessResponse'
            - properties:
                ' Message':
                  type: string
                Data:
                  properties:
                    tournament_id:
                      type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create tournament
      tags:
      - tournaments
  /tournaments/{id}:
    get:
      description: Returns a single tournament by ID
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.SuccessResponse'
            - properties:
                ' Message':
                  type: string
                Data:
                  properties:
                    tournament:
                      $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.Tournament'
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get tournament
      tags:
      - tournaments
  /tournaments/{id}/groups/{groupId}/standings:
    get:
      description: Returns standings for a single group with the full tiebreaker cascade
        applied (head-to-head → mutual set ratio → mutual game ratio → overall set
        ratio → overall game ratio → draw). Each entry includes a rank field.
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: string
      - description: Group ID
        in: path
        name: groupId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.SuccessResponse'
            - properties:
                ' Message':
                  type: string
                Data:
                  properties:
                    Standings:
                      items:
                        $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.Standings'
                      type: array
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get group standings
      tags:
      - standings
  /tournaments/{id}/matches/{matchId}:
    get:
      description: Returns a match with its full participant list and set-by-set scores
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: string
      - description: Match ID
        in: path
        name: matchId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.SuccessResponse'
            - properties:
                ' Message':
                  type: string
                Data:
                  properties:
                    match:
                      $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.MatchDetail'
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get match detail
      tags:
      - matches
  /tournaments/{id}/matches/{matchId}/result:
    post:
      consumes:
      - application/json
      description: Records the full set-by-set result for a match. Validates set count
        against best_of, updates standings for group/round-robin stages, and advances
        the winner to the next bracket slot atomically.
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: string
      - description: Match ID
        in: path
        name: matchId
        required: true
        type: string
      - description: Set scores
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.SubmitResultRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.SuccessResponse'
            - properties:
                ' Message':
                  type: string
                Data:
                  properties:
                    result_detail:
                      $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.MatchDetail'
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit match result
      tags:
      - matches
    put:
      consumes:
      - application/json
      description: Voids the previous result, reverses standings contributions, and
        resubmits with corrected scores. Blocked if the next match in the bracket
        has already been completed. Admin only.
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: string
      - description: Match ID
        in: path
        name: matchId
        required: true
        type: string
      - description: Corrected set scores
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.SubmitResultRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.SuccessResponse'
            - properties:
                ' Message':
                  type: string
                Data:
                  properties:
                    result_detail:
                      $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.MatchDetail'
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Correct match result
      tags:
      - matches
...
  /tournaments/{id}/stages/{stageId}/matches:
    get:
      description: Returns all matches for a stage ordered by round and position
      parameters:
      - description: Tournament ID
        in: path
        name: id
        required: true
        type: string
      - description: Stage ID
        in: path
        name: stageId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.SuccessResponse'
            - properties:
                ' Message':
                  type: string
                Data:
                  properties:
                    matches:
                      items:
                        $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.Match'
                      type: array
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uthmanduro_BracketForge_internal_model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List stage matches
      tags:
      - matches
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"